Swapping input test approach for coverage
